+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## dev ##
 
 ### enhancements
+  * add CityHash.hash32 method
   * update libcityhash to 1.1. See [NEWS](http://code.google.com/p/cityhash/source/browse/trunk/NEWS)